% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbr2state.R
\name{abbr2state}
\alias{abbr2state}
\title{Convert state abbreviations to names}
\usage{
abbr2state(abbr)
}
\arguments{
\item{abbr}{A vector of state abbreviation.}
}
\value{
Returns a vector of the same length with the corresponding state
names or abbreviations.
}
\description{
Two utility functions. One converts state names to the state abbreviations,
and the second does the opposite.
}
\examples{

abbr2state("MN")

}
\seealso{
\code{\link{state2abbr}}, \code{\link{county}}, \code{\link{county_complete}}
}
\author{
David Diez
}
\keyword{Abbreviation}
\keyword{State}
