\name{association.4.6}
\alias{association.4.6}
\docType{data}
\title{Simulated data for association plots, set 2}
\description{Simulated data set.}
\usage{data("association.4.6")}
\format{
  A data frame with 121 observations on the following 4 variables.
  \describe{
    \item{\code{x2}}{a numeric vector}
    \item{\code{y4}}{a numeric vector}
    \item{\code{y5}}{a numeric vector}
    \item{\code{y6}}{a numeric vector}
  }
}
\examples{
data(association.4.6)
plot(association.4.6$x2, association.4.6$y4)
}
\keyword{datasets}
