\name{association.7.12}
\alias{association.7.12}
\docType{data}
\title{Simulated data for association plots, set 3}
\description{Simulated data set.}
\usage{data("association.7.12")}
\format{
  A data frame with 141 observations on the following 7 variables.
  \describe{
    \item{\code{x3}}{a numeric vector}
    \item{\code{y7}}{a numeric vector}
    \item{\code{y8}}{a numeric vector}
    \item{\code{y9}}{a numeric vector}
    \item{\code{y10}}{a numeric vector}
    \item{\code{y11}}{a numeric vector}
    \item{\code{y12}}{a numeric vector}
  }
}
\examples{
data(association.7.12)
plot(association.7.12$x3, association.7.12$y7)
}
\keyword{datasets}
