% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-association.R
\docType{data}
\name{association}
\alias{association}
\alias{association_1_3,}
\alias{association_4_6,}
\alias{association_7_12}
\title{Simulated data for association plots}
\format{
A data frame with 121 observations on the following 4 variables.
\describe{
\item{x1}{a numeric vector}
\item{x2}{a numeric vector}
\item{x3}{a numeric vector}
\item{y1}{a numeric vector}
\item{y2}{a numeric vector}
\item{y3}{a numeric vector}
\item{y4}{a numeric vector}
\item{y5}{a numeric vector}
\item{y6}{a numeric vector}
\item{y7}{a numeric vector}
\item{y8}{a numeric vector}
\item{y9}{a numeric vector}
\item{y10}{a numeric vector}
\item{y11}{a numeric vector}
\item{y12}{a numeric vector}
}
}
\usage{
association
}
\description{
Simulated dataset.
}
\examples{

library(ggplot2)

ggplot(association, aes(x = x1, y = y1)) +
  geom_point()

ggplot(association, aes(x = x2, y = y4)) +
  geom_point()

ggplot(association, aes(x = x3, y = y7)) +
  geom_point()
}
\keyword{datasets}
