% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-avandia.R
\docType{data}
\name{avandia}
\alias{avandia}
\title{Cardiovascular problems for two types of Diabetes medicines}
\format{
A data frame with 227571 observations on the following 2 variables.
\describe{
\item{treatment}{a factor with levels \code{Pioglitazone} and
\code{Rosiglitazone}}
\item{cardiovascular_problems}{a factor with levels \code{no} and
\code{yes}}
}
}
\source{
D.J. Graham et al. Risk of acute myocardial infarction, stroke,
heart failure, and death in elderly Medicare patients treated with
rosiglitazone or pioglitazone. In: JAMA 304.4 (2010), p. 411. issn:
0098-7484.
}
\usage{
avandia
}
\description{
A comparison of cardiovascular problems for Rosiglitazone and Pioglitazone.
}
\examples{

table(avandia)

}
\keyword{datasets}
