% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-baby_names.R
\docType{data}
\name{baby_names}
\alias{baby_names}
\title{Baby Names Data.}
\format{
A data frame with 2820 rows and 4 variables.
\describe{
\item{rank}{Rank of name in year.}
\item{male}{Boy's name.}
\item{female}{Girl's name.}
\item{year}{Year of name rank.}
}
}
\source{
\href{https://www.ssa.gov/OACT/babynames/background.html}{Social Security Administration}
}
\usage{
baby_names
}
\description{
From USA Social Security Administration, baby names from 1880 to 2020,
the top 20 for each year
}
\examples{
library(dplyr)

# List the 10 most frequent male names
baby_names \%>\% 
  group_by(male) \%>\%
  summarise(freq=n()) \%>\%
  arrange(desc(freq)) \%>\% 
  top_n(n = 10)

# List Female names in top 20 in both 1880 and 1980
baby_names \%>\% 
select(female,year) \%>\%  
  group_by(female) \%>\%
  filter(year==1880 | year == 1980) \%>\%
  filter(n() > 1) \%>\%
  select(female) \%>\%
  distinct()

}
\keyword{datasets}
