% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BG.R
\name{BG}
\alias{BG}
\title{Overlay a color on an entire plotting window}
\usage{
BG(col = openintro::COL[5, 9])
}
\arguments{
\item{col}{Color to be overlayed}
}
\description{
For overlaying a color on an entire plotting window
}
