% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_streak.R
\name{calc_streak}
\alias{calc_streak}
\title{Calculate hit streaks}
\usage{
calc_streak(x)
}
\arguments{
\item{x}{A character vector of hits (\code{"H"}) and misses (\code{"M"}).}
}
\value{
A data frame with one column, \code{length}, containing the length of
each hit streak.
}
\description{
Calculate hit streaks
}
\examples{
data(kobe_basket)
calc_streak(kobe_basket$shot)
}
