% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cancer_in_dogs.R
\docType{data}
\name{cancer_in_dogs}
\alias{cancer_in_dogs}
\title{Cancer in dogs}
\format{
A data frame with 1436 observations on the following 2 variables.
\describe{
\item{order}{a factor with levels \code{2,4-D} and \code{no 2,4-D}}
\item{response}{a factor with levels \code{cancer} and \code{no cancer}}
}
}
\source{
Hayes HM, Tarone RE, Cantor KP, Jessen CR, McCurnin DM, and
Richardson RC. 1991. Case- Control Study of Canine Malignant Lymphoma:
Positive Association With Dog Owner's Use of 2, 4- Dichlorophenoxyacetic
Acid Herbicides. Journal of the National Cancer Institute 83(17):1226-1231.
}
\usage{
cancer_in_dogs
}
\description{
A study in 1994 examined 491 dogs that had developed cancer and 945 dogs as
a control group to determine whether there is an increased risk of cancer in
dogs that are exposed to the herbicide 2,4-Dichlorophenoxyacetic acid
(2,4-D).
}
\examples{

table(cancer_in_dogs)

}
\keyword{datasets}
