% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-chickwts.R
\docType{data}
\name{chickwts}
\alias{chickwts}
\alias{chick.wts}
\title{Chicken weights vs feed type}
\format{A tibble with 71 observations on the following 2 variables.
\describe{
  \item{weight}{a numeric vector}
  \item{feed}{a factor with levels \code{casein} \code{horsebean}
  \code{linseed} \code{meatmeal} \code{soybean} \code{sunflower}}
}}
\source{
Chicken Weights by Feed Type, from the \code{datasets} package in R.

McNeil, D. R. (1977) Interactive Data Analysis. New York: Wiley.
}
\usage{
chickwts
}
\description{
Experiment for chick weights and feed type.
}
\examples{

boxplot(chickwts$weight ~ chickwts$feed)

}
\keyword{datasets}
