% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-coast_starlight.R
\docType{data}
\name{coast_starlight}
\alias{coast_starlight}
\alias{coast.starlight}
\title{Coast Starlight Amtrak train}
\format{
A data frame with 16 observations on the following 3 variables.
\describe{
\item{station}{Station.}
\item{dist}{Distance.}
\item{travel_time}{Travel time.}
}
}
\usage{
coast_starlight
}
\description{
Travel times and distances.
}
\examples{

library(ggplot2)

ggplot(coast_starlight, aes(x = dist, y = travel_time)) +
  geom_point()

}
\keyword{datasets}
