% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-corr_match.R
\docType{data}
\name{corr_match}
\alias{corr_match}
\alias{corr.match}
\title{Sample data sets for correlation problems}
\format{
A data frame with 121 observations on the following 9 variables.
\describe{
  \item{x}{a numeric vector}
  \item{y1}{a numeric vector}
  \item{y2}{a numeric vector}
  \item{y3}{a numeric vector}
  \item{y4}{a numeric vector}
  \item{y5}{a numeric vector}
  \item{y6}{a numeric vector}
  \item{y7}{a numeric vector}
  \item{y8}{a numeric vector}
  }
}
\source{
Simulated data set.
}
\usage{
corr_match
}
\description{
Simulated data.
}
\examples{

library(ggplot2)

ggplot(corr_match, aes(x = x, y = y1)) +
  geom_point()

cor(corr_match$x, corr_match$y1)

}
\keyword{datasets}
