% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-county_2023.R
\docType{data}
\name{county_2023}
\alias{county_2023}
\title{county_2023}
\format{
A data frame with 3144 rows and 43 columns.
\describe{
\item{\code{geoid}}{Unique numeric code used by the U.S. Census Bureau to identify specific geographic areas.}
\item{\code{county_state}}{Name of county and state.}
\item{\code{name}}{Name of county.}
\item{\code{state}}{Name of state.}
\item{\code{individual_income}}{Median individual income for the county, measure in 2018 USD.}
\item{\code{population_size}}{Number of people residing in the county.}
\item{\code{white}}{Percent of the population who identified as only white.}
\item{\code{african_american}}{Percent of the population who identified as only Black or African American.}
\item{\code{age}}{Median age of those who live in the county, measured in years.}
\item{\code{age_under_5}}{Percent of the population under the age of 5.}
\item{\code{age_85_plus}}{Percent of the population age 85 and older.}
\item{\code{age_18_plus}}{Percent of the population age 18 and older.}
\item{\code{age_65_plus}}{Percent of the population age 65 and older.}
\item{\code{travel_time}}{Mean travel time to work, measured in minutes.}
\item{\code{household_size}}{Mean household size.}
\item{\code{family_size}}{Mean family size.}
\item{\code{one_unit_structure}}{Percent of properties with a residential building that contains a single housing dwelling unit.}
\item{\code{two_unit_structure}}{Percent of properties with a residential building that contains two or more separate housing units (such as a duplex or apartments).}
\item{\code{mobile_and_other}}{Percent of properties with mobile homes and all other types of housing structures.}
\item{\code{individual_income_25_plus}}{Median individual income for those age 25 or older.}
\item{\code{hs_grad}}{Percent of the population over the age 25 or older with at least a high school diploma.}
\item{\code{college_grad}}{Percent of the population over the age 25 or older with at least a Bachelor's Degree.}
\item{\code{total_households}}{Total number of households in the county.}
\item{\code{spanish_speaking}}{Percent of households that speak Spanish.}
\item{\code{indo_european_language}}{Percent of households that speak an Indo-European language other than English.}
\item{\code{api_language}}{Percent of households that speak an Asian and Pacific Islander language.}
\item{\code{non_euro_api}}{Percent of households speaking an non-European or Asian and Pacific Islander language.}
\item{\code{limited_english}}{Percent of households that speak limited English.}
\item{\code{poverty}}{Percent of population living below the poverty line.}
\item{\code{poverty_18}}{Percent of population under the age of 18 living below the poverty line.}
\item{\code{poverty_65}}{Percent of the population over the age of 65 living below the povery line.}
\item{\code{household_income_mean}}{Mean household income, measured in 2018 USD.}
\item{\code{household_income_median}}{Median household income, measured in 2018 USD.}
\item{\code{per_capita_income}}{Mean income per capita, measured in 2018 USD.}
\item{\code{veterans}}{Percent of the civilian population age 18 or older who served in the military.}
\item{\code{unemployment_20_64}}{Unemployment rate for those age 20 to 64.}
\item{\code{uninsured}}{Percent of the civilian population who is uninsured.}
\item{\code{uninsured_6}}{Percent of population under the age of 6 who are uninsured.}
\item{\code{uninsured_19}}{Percent of the population under the age of 19 who are uninsured.}
\item{\code{uninsured_65}}{Percent of the population age 65 and older who are uninsured.}
\item{\code{computers}}{Percent of households that have a laptop or desktop computer.}
\item{\code{smartphones}}{Percent of households that have a smartphone.}
\item{\code{broadband}}{Percent of households with computers that have access to broadband internet.}
}
}
\source{
\url{https://www.census.gov/programs-surveys/acs/data.html} collected using the \code{tidycensus} package.
}
\usage{
county_2023
}
\description{
County level data from the US Census American Community Survey 2023 5-year estimates.
}
\keyword{datasets}
