% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-drug_use.R
\docType{data}
\name{drug_use}
\alias{drug_use}
\title{Drug use of students and parents}
\format{
A data frame with 445 observations on the following 2 variables.
\describe{
\item{student}{a factor with levels \code{not} \code{uses}}
\item{parents}{a factor with levels \code{not} \code{used}}
}
}
\source{
Ellis GJ and Stone LH. 1979. Marijuana Use in College: An
Evaluation of a Modeling Explanation. Youth and Society 10:323-334.
}
\usage{
drug_use
}
\description{
Summary of 445 student-parent pairs.
}
\examples{

table(drug_use)
}
\keyword{datasets}
