% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edaPlot.R
\name{edaPlot}
\alias{edaPlot}
\alias{createEdaOptions}
\alias{guessMethod}
\alias{plotNothing}
\alias{fitNormal}
\alias{makePlotIcon}
\title{Exploratory data analysis plot}
\usage{
edaPlot(
  dataFrame,
  Col = c("#888888", "#FF0000", "#222222", "#FFFFFF", "#CCCCCC", "#3377AA")
)
}
\arguments{
\item{dataFrame}{A data frame.}

\item{Col}{A vector containing six colors. The colors may be given in any
form.}
}
\description{
Explore different plotting methods using a click interface.
}
\examples{

data(mlbbat10)
bat <- mlbbat10[mlbbat10$at_bat > 200,]
#edaPlot(bat)

data(mariokart)
mk <- mariokart[mariokart$total_pr < 100,]
#edaPlot(mk)

}
\seealso{
\code{\link{histPlot}}, \code{\link{densityPlot}},
\code{\link{boxPlot}}, \code{\link{dotPlot}}
}
\author{
David Diez
}
