% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-EVs_region.R
\docType{data}
\name{EVs_region}
\alias{EVs_region}
\title{Electric vehicles}
\format{
A data frame with 5085 observations on the following 2 variables.
\describe{
\item{\code{region}}{a factor with levels \code{Urban, Suburban} and \code{Rural}}
\item{\code{consideration}}{a factor with levels \code{Not too or not at all likely}, \code{Very or somewhat likely} and \code{Do not expect to buy a vehicle}}
}
}
\source{
Pew Research Center, Survey of U.S. adults conducted April 28-May 4, 2025. "Americans' Views on Energy at the Start of Trump's Second Term"
}
\usage{
EVs_region
}
\description{
Pew research results regarding consideration of buying electric vehicle.
}
\keyword{datasets}
