% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-exclusive_relationship.R
\docType{data}
\name{exclusive_relationship}
\alias{exclusive_relationship}
\alias{exclusive.relationship}
\title{Number of Exclusive Relationships}
\format{
A data frame with 218 observations on the following variable.
\describe{\item{num}{Number of exclusive relationships.} }
}
\usage{
exclusive_relationship
}
\description{
A survey conducted on a reasonably random sample of 203 undergraduates
asked, among many other questions, about the number of exclusive
relationships these students have been in.
}
\examples{

summary(exclusive_relationship$num)
table(exclusive_relationship$num)
hist(exclusive_relationship$num)
}
\keyword{datasets}
