% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-famuss.R
\docType{data}
\name{famuss}
\alias{famuss}
\title{A dataset to examine the relationship between muscle strength and the single nucleotide polymorphism (SNP) actn3.r577x.}
\format{
A tibble with 1397 rows and 10 variables
\describe{
\item{\code{ndrm.ch}}{A numeric vector, the percent change in strength
in a non-dominant arm, from before training and after.}
\item{\code{drm.ch}}{A numeric vector, percent change in strength in
dominant arm.}
\item{\code{sex}}{A factor with levels \code{Female} and \code{Male}}
\item{\code{age}}{A numeric vector, age in years.}
\item{\code{race}}{A factor with levels \code{African Am} \code{Asian}
\code{Caucasian} \code{Hispanic} \code{Other}}
\item{\code{height}}{A numeric vector,
height in inches.}
\item{\code{weight}}{A numeric vector, weight in pounds.}
\item{\code{actn3.r577x}}{A factor with levels \code{CC} \code{CT} \code{TT},
that shows the genotype at residue rs540874 (location r577x) within the ACTN3
SNP.}
\item{\code{bmi}}{A numeric vector, body mass index}
}
}
\source{
Personal communication from A. Foulkes
}
\usage{
famuss
}
\description{
This dataset is a  subset of the larger data set from the Functional SNPs
Associated with Muscle Size and Strength (FAMuSS) by Thompson et.al. It
contains demographic, response and coding for the SNP for the study participants.
Unlike the data in the previous version of the \code{oibiostat} data package,
this dataset retains the missing values. The data are also discussed in the
Foulkes text. Strength was measured in both dominant and non-dominant arms
before and after resistance training. The particular gene of interest was
ACTN3, the "sports gene."
}
\references{
Thompson PMoyna NSeip R et al. Medicine and Science in Sports and
Exercise, (2004), 1132-1139, 36(7). Clarkson P, et al., Journal of Applied
Physiology 99: 154-163, 2005.Pescatello L, et al. Highlights from the
functional single nucleotide polymorphisms associated with human muscle
size and strength or FAMuSS study, BioMed Research International 2013. Foulkes, Andrea S.
Applied Statistical Genetics using R for Population Association Studies.
Springer, 2009).
}
\keyword{datasets}
