% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fastfood.R
\docType{data}
\name{fastfood}
\alias{fastfood}
\title{Nutrition in fast food}
\format{
A data frame with 515 observations on the following 17 variables.
\describe{
\item{restaurant}{Name of restaurant}
\item{item}{Name of item}
\item{calories}{Number of calories}
\item{cal_fat}{Calories from fat}
\item{total_fat}{Total fat}
\item{sat_fat}{Saturated fat}
\item{trans_fat}{Trans fat}
\item{cholesterol}{Cholesterol}
\item{sodium}{Sodium}
\item{total_carb}{Total carbs}
\item{fiber}{Fiber}
\item{sugar}{Suger}
\item{protein}{Protein}
\item{vit_a}{Vitamin A}
\item{vit_c}{Vitamin C}
\item{calcium}{Calcium}
\item{salad}{Salad or not}
}
}
\source{
Retrieved from \href{https://github.com/rfordatascience/tidytuesday/tree/master/data/2018/2018-09-04}{Tidy Tuesday Fast food entree data}.
}
\usage{
fastfood
}
\description{
Nutrition amounts in 515 fast food items.  The author of the data
scraped only entrees (not sides, drinks, desserts, etc.).
}
\keyword{datasets}
