\name{fcid}
\alias{fcid}
\docType{data}
\title{Summary of male heights from USDA Food Commodity Intake Database}
\description{Sample of heights based on the weighted sample in the survey.}
\usage{data("fcid")}
\format{
  A data frame with 100 observations on the following 2 variables.
  \describe{
    \item{\code{height}}{a numeric vector}
    \item{\code{num_of_adults}}{a numeric vector}
  }
}
\examples{
data(fcid)
}
\keyword{datasets}
