% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fheights.R
\docType{data}
\name{fheights}
\alias{fheights}
\title{Female college student heights, in inches}
\format{
A data frame with 24 observations on the following variable.
\describe{\item{heights}{height, in inches} }
}
\usage{
fheights
}
\description{
24 sample observations.
}
\examples{

hist(fheights$heights)
}
\keyword{datasets}
