% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-full_body_scan.R
\docType{data}
\name{full_body_scan}
\alias{full_body_scan}
\alias{full.body.scan}
\title{Poll about use of full-body airport scanners}
\format{
A data frame with 1137 observations on the following 2 variables.
\describe{\item{answer}{a factor with levels \code{do not know / no
answer} \code{should} \code{should not}} \item{party.affiliation}{a
factor with levels \code{Democrat} \code{Independent} \code{Republican}} }
}
\source{
S. Condon. Poll: 4 in 5 Support Full-Body Airport Scanners. In: CBS
News (2010).
}
\usage{
full_body_scan
}
\description{
Poll about use of full-body airport scanners, where about 4-in-5 people
supported the use of the scanners.
}
\examples{

full_body_scan
}
\keyword{datasets}
