% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gear_company.R
\docType{data}
\name{gear_company}
\alias{gear_company}
\title{Fake data for a gear company example}
\format{
A data frame with 2000 observations on the following 2 variables.
\describe{\item{company}{a factor with levels \code{current}
\code{prospective}} \item{outcome}{a factor with levels \code{not}
\code{pass}} }
}
\usage{
gear_company
}
\description{
Made-up data for whether a sample of two gear companies' parts pass
inspection.
}
\examples{

gear_company
}
\keyword{datasets}
