% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-global_warming_pew.R
\docType{data}
\name{global_warming_pew}
\alias{global_warming_pew}
\alias{global.warming.pew}
\title{Pew survey on global warming}
\format{
A data frame with 2253 observations on the following 2 variables.
\describe{\item{party_or_ideology}{a factor with levels
\code{Conservative Republican} \code{Liberal Democrat} \code{Mod/Cons
Democrat} \code{Mod/Lib Republican}} \item{response}{Response.} }
}
\source{
Pew Research Center, Majority of Republicans No Longer See Evidence
of Global Warming, data collected on October 27, 2010.
}
\usage{
global_warming_pew
}
\description{
A 2010 Pew Research poll asked 1,306 Americans, "From what you've read and
heard, is there solid evidence that the average temperature on earth has
been getting warmer over the past few decades, or not?"
}
\examples{

global_warming_pew
}
\keyword{datasets}
