% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gov_poll.R
\docType{data}
\name{gov_poll}
\alias{gov_poll}
\title{Pew Research poll on government approval ratings}
\format{
A data frame with 4223 observations on the following 2 variables.
\describe{\item{poll}{a factor with levels \code{approve}
\code{disapprove}} \item{eval}{a factor with levels \code{Democrats}
\code{Obama} \code{Republicans}} }
}
\source{
See the Pew Research website:
www.people-press.org/2012/03/14/romney-leads-gop-contest-trails-in-
matchup-with-obama. The counts in Table 6.19 are approximate.
}
\usage{
gov_poll
}
\description{
The poll's focus is on Obama and then Democrats and Republicans in Congress.
}
\examples{

gov_poll
}
\keyword{datasets}
