\name{gpa.iq}
\alias{gpa.iq}
\alias{gpa_iq}
\docType{data}
\title{
  Sample of students and their GPA and IQ
}
\description{
  Data on 78 students including GPA, IQ, and gender.
}
\usage{data("gpa.iq")}
\format{
  A data frame with 78 observations representing students on the following 5 variables.
  \describe{
    \item{\code{obs}}{a numeric vector}
    \item{\code{gpa}}{Grade point average (GPA).}
    \item{\code{iq}}{IQ.}
    \item{\code{gender}}{Gender.}
    \item{\code{concept}}{a numeric vector}
  }
}
\examples{
data(gpa.iq)
}
\keyword{datasets}
