% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gpa_iq.R
\docType{data}
\name{gpa_iq}
\alias{gpa_iq}
\alias{gpa.iq}
\title{Sample of students and their GPA and IQ}
\format{
A data frame with 78 observations representing students on the
following 5 variables. \describe{\item{obs}{a numeric vector}
\item{gpa}{Grade point average (GPA).} \item{iq}{IQ.}
\item{gender}{Gender.} \item{concept}{a numeric vector} }
}
\usage{
gpa_iq
}
\description{
Data on 78 students including GPA, IQ, and gender.
}
\examples{

gpa_iq

}
\keyword{datasets}
