\name{gss2010}
\alias{gss2010}
\docType{data}
\title{2010 General Social Survey}
\description{
  Data from the 2010 General Social Survey.
}
\usage{data("gss2010")}
\format{
  A data frame with 2044 observations on the following 5 variables.
  \describe{
    \item{\code{hrsrelax}}{
      After an average work day, about how many hours
      do you have to relax or pursue activities that
      you enjoy?
    }
    \item{\code{mntlhlth}}{
      For how many days during the past 30 days was your
      mental health, which includes stress, depression,
      and problems with emotions, not good?
    }
    \item{\code{hrs1}}{Hours worked each week.}
    \item{\code{degree}}{Educational attainment or degree.}
    \item{\code{grass}}{
      Do you think the use of marijuana should be
      made legal, or not?
    }
  }
}
\source{US 2010 General Social Survey.}
\examples{
data(gss2010)
}
\keyword{datasets}
