% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gss_wordsum_class.R
\docType{data}
\name{gss_wordsum_class}
\alias{gss_wordsum_class}
\title{gss_wordsum_class}
\format{
A data frame with 795 observations on the following 2 variables.
\describe{
\item{wordsum}{A vocabulary score calculated based on a ten question vocabulary test, where a higher score means better vocabulary. Scores range from 1 to 10.}
\item{class}{Self-identified social class has 4 levels: lower, working, middle, and upper class.}
}
}
\usage{
gss_wordsum_class
}
\description{
A data frame containing data from the General Social Survey.
}
\examples{

library(dplyr)

gss_wordsum_class |>
  group_by(class) |>
  summarize(mean_wordsum = mean(wordsum))

}
\keyword{datasets}
