% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-health_coverage.R
\docType{data}
\name{health_coverage}
\alias{health_coverage}
\alias{health.coverage}
\title{Health Coverage and Health Status}
\format{
A data frame with 20000 observations on the following 2 variables.
\describe{\item{coverage}{Whether the person had health coverage or not.}
\item{health_status}{The person's health status.} }
}
\source{
Office of Surveillance, Epidemiology, and Laboratory Services
Behavioral Risk Factor Surveillance System, BRFSS 2010 Survey Data.
}
\usage{
health_coverage
}
\description{
Survey responses for 20,000 responses to the Behavioral Risk Factor
Surveillance System.
}
\examples{


table(health_coverage)
}
\keyword{datasets}
