% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-hfi.R
\docType{data}
\name{hfi}
\alias{hfi}
\title{Absenteeism}
\format{A tibble with `1458` observations on the following 123 variables.
\describe{
  \item{year}{}
  \item{ISO_code}{}
  \item{countries}{}
  \item{region}{}
  \item{pf_rol_procedural}{}
  \item{pf_rol_civil}{}
  \item{pf_rol_criminal}{}
  \item{pf_rol}{}
  \item{pf_ss_homicide}{}
  \item{pf_ss_disappearances_disap}{}
  \item{pf_ss_disappearances_violent}{}
  \item{pf_ss_disappearances_organized}{}
  \item{pf_ss_disappearances_fatalities}{}
  \item{pf_ss_disappearances_injuries}{}
  \item{pf_ss_disappearances}{}
  \item{pf_ss_women_fgm}{}
  \item{pf_ss_women_missing}{}
  \item{pf_ss_women_inheritance_widows}{}
  \item{pf_ss_women_inheritance_daughters}{}
  \item{pf_ss_women_inheritance}{}
  \item{pf_ss_women}{}
  \item{pf_ss}{}
  \item{pf_movement_domestic}{}
  \item{pf_movement_foreign}{}
  \item{pf_movement_women}{}
  \item{pf_movement}{}
  \item{pf_religion_estop_establish}{}
  \item{pf_religion_estop_operate}{}
  \item{pf_religion_estop}{}
  \item{pf_religion_harassment}{}
  \item{pf_religion_restrictions}{}
  \item{pf_religion}{}
  \item{pf_association_association}{}
  \item{pf_association_assembly}{}
  \item{pf_association_political_establish}{}
  \item{pf_association_political_operate}{}
  \item{pf_association_political}{}
  \item{pf_association_prof_establish}{}
  \item{pf_association_prof_operate}{}
  \item{pf_association_prof}{}
  \item{pf_association_sport_establish}{}
  \item{pf_association_sport_operate}{}
  \item{pf_association_sport}{}
  \item{pf_association}{}
  \item{pf_expression_killed}{}
  \item{pf_expression_jailed}{}
  \item{pf_expression_influence}{}
  \item{pf_expression_control}{}
  \item{pf_expression_cable}{}
  \item{pf_expression_newspapers}{}
  \item{pf_expression_internet}{}
  \item{pf_expression}{}
  \item{pf_identity_legal}{}
  \item{pf_identity_parental_marriage}{}
  \item{pf_identity_parental_divorce}{}
  \item{pf_identity_parental}{}
  \item{pf_identity_sex_male}{}
  \item{pf_identity_sex_female}{}
  \item{pf_identity_sex}{}
  \item{pf_identity_divorce}{}
  \item{pf_identity}{}
  \item{pf_score}{}
  \item{pf_rank}{}
  \item{ef_government_consumption}{}
  \item{ef_government_transfers}{}
  \item{ef_government_enterprises}{}
  \item{ef_government_tax_income}{}
  \item{ef_government_tax_payroll}{}
  \item{ef_government_tax}{}
  \item{ef_government}{}
  \item{ef_legal_judicial}{}
  \item{ef_legal_courts}{}
  \item{ef_legal_protection}{}
  \item{ef_legal_military}{}
  \item{ef_legal_integrity}{}
  \item{ef_legal_enforcement}{}
  \item{ef_legal_restrictions}{}
  \item{ef_legal_police}{}
  \item{ef_legal_crime}{}
  \item{ef_legal_gender}{}
  \item{ef_legal}{}
  \item{ef_money_growth}{}
  \item{ef_money_sd}{}
  \item{ef_money_inflation}{}
  \item{ef_money_currency}{}
  \item{ef_money}{}
  \item{ef_trade_tariffs_revenue}{}
  \item{ef_trade_tariffs_mean}{}
  \item{ef_trade_tariffs_sd}{}
  \item{ef_trade_tariffs}{}
  \item{ef_trade_regulatory_nontariff}{}
  \item{ef_trade_regulatory_compliance}{}
  \item{ef_trade_regulatory}{}
  \item{ef_trade_black}{}
  \item{ef_trade_movement_foreign}{}
  \item{ef_trade_movement_capital}{}
  \item{ef_trade_movement_visit}{}
  \item{ef_trade_movement}{}
  \item{ef_trade}{}
  \item{ef_regulation_credit_ownership}{}
  \item{ef_regulation_credit_private}{}
  \item{ef_regulation_credit_interest}{}
  \item{ef_regulation_credit}{}
  \item{ef_regulation_labor_minwage}{}
  \item{ef_regulation_labor_firing}{}
  \item{ef_regulation_labor_bargain}{}
  \item{ef_regulation_labor_hours}{}
  \item{ef_regulation_labor_dismissal}{}
  \item{ef_regulation_labor_conscription}{}
  \item{ef_regulation_labor}{}
  \item{ef_regulation_business_adm}{}
  \item{ef_regulation_business_bureaucracy}{}
  \item{ef_regulation_business_start}{}
  \item{ef_regulation_business_bribes}{}
  \item{ef_regulation_business_licensing}{}
  \item{ef_regulation_business_compliance}{}
  \item{ef_regulation_business}{}
  \item{ef_regulation}{}
  \item{ef_score}{}
  \item{ef_rank}{}
  \item{hf_score}{}
  \item{hf_rank}{}
  \item{hf_quartile}{}
  }}
\usage{
hfi
}
\description{
The Human Freedom Index is a report that attempts to summarize the idea of "freedom"
through a bunch of different variables for many countries around the globe.
It serves as a rough objective measure for the relationships between the different
types of freedom - whether it's political, religious, economical or personal freedom -
and other social and economic circumstances. The Human Freedom Index is an anually
co-published report by the Cato Institute, the Fraser Institute, and the Liberales
Institut at the Friedrich Naumann Fourndation for Freedom.
}
\details{
This dataset contains information from Human Freedom Index reports from 2008-2016.
}
\keyword{datasets}
