\name{housing}
\alias{housing}
\docType{data}
\title{Simulated data set on student housing}
\description{Each observation represents a simulated rent price for a student.}
\usage{data("housing")}
\format{
  A data frame with 75 observations on the following variable.
  \describe{
    \item{\code{cost}}{a numeric vector}
  }
}
\examples{
data(housing)
## maybe str(housing) ; plot(housing) ...
}
\keyword{datasets}
