\name{jury}
\alias{jury}
\docType{data}
\title{Simulated juror data set}
\description{Simulated data set of registered voters proportions and representation on juries.}
\usage{data("jury")}
\format{
  A data frame with 275 observations on the following variable.
  \describe{
    \item{\code{race}}{a factor with levels \code{black} \code{hispanic} \code{other} \code{white}}
  }
}
\examples{
data(jury)
## maybe str(jury) ; plot(jury) ...
}
\keyword{datasets}