% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-kobe_basket.R
\docType{data}
\name{kobe_basket}
\alias{kobe_basket}
\title{Kobe Bryant basketball performance}
\format{
A data frame with 133 rows and 6 variables:
\describe{
  \item{vs}{A categorical vector, ORL if the Los Angeles Lakers played
  against Orlando}
  \item{game}{A numerical vector, game in the 2009 NBA finals}
  \item{quarter}{A categorical vector, quarter in the game, OT stands for
  overtime}
  \item{time}{A character vector, time at which Kobe took a shot}
  \item{description}{A character vector, description of the shot}
  \item{shot}{A categorical vector, H if the shot was a hit, M if the shot
  was a miss}
}
}
\usage{
kobe_basket
}
\description{
Data from the five games the Los Angeles Lakers played against the Orlando
Magic in the 2009 NBA finals.
}
\details{
Each row represents a shot Kobe Bryant took during the five games of the
2009 NBA finals. Kobe Bryant's performance earned him the title of Most
Valuable Player and many spectators commented on how he appeared to show
a hot hand.
}
\keyword{datasets}
