% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-leg_mari.R
\docType{data}
\name{leg_mari}
\alias{leg_mari}
\title{Legalization of Marijuana Support in 2010 California Survey}
\format{
A data frame with 119 observations on the following variable.
\describe{\item{response}{One of two values: \code{oppose} or
\code{support}.} }
}
\source{
Survey USA, Election Poll #16804, data collected July 8-11, 2010.
}
\usage{
leg_mari
}
\description{
In a 2010 Survey USA poll, 70% of the 119 respondents between the ages of 18
and 34 said they would vote in the 2010 general election for Prop 19, which
would change California law to legalize marijuana and allow it to be
regulated and taxed.
}
\examples{


table(leg_mari)


}
\keyword{datasets}
