% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-life_exp.R
\docType{data}
\name{life_exp}
\alias{life_exp}
\title{life_exp}
\format{
A data frame with 3142 observations on the following 4 variables.
\describe{
\item{state}{Name of the state.}
\item{county}{Name of the county.}
\item{expectancy}{Life expectancy in the county.}
\item{income}{Median income in the county, measured in US $.}
}
}
\usage{
life_exp
}
\description{
A data frame with 3142 rows and 4 columns. County level data for life expectancy and median income in the United States.
}
\examples{

library(ggplot2)

# Income V Expectancy
ggplot(life_exp, aes(x = income, y = expectancy)) +
  geom_point(color = openintro::IMSCOL["green", "full"], alpha = 0.2) +
  theme_minimal() +
  labs(
    title = "Is there a relationship between median income and life expectancy?",
    x = "Median income (US $)",
    y = "Life Expectancy (year)"
  )
}
\keyword{datasets}
