\name{male.heights}
\alias{male.heights}
\docType{data}
\title{Sample of 100 male heights}
\description{Random sample based on Food Commodity Intake Database distribution}
\usage{data("male.heights")}
\format{
  A data frame with 100 observations on the following variable.
  \describe{
    \item{\code{heights}}{a numeric vector}
  }
}
\references{http://fcid.foodrisk.org/}
\examples{
data(male.heights)
## maybe str(male.heights) ; plot(male.heights) ...
}
\keyword{datasets}