% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-marathon.R
\docType{data}
\name{marathon}
\alias{marathon}
\title{New York City Marathon Times (outdated)}
\format{
A data frame with 60 observations on the following 3 variables.
\describe{
\item{year}{Year}
\item{gender}{Gender}
\item{time}{Running time (in hours)}
}
}
\source{
Data source has been removed.
}
\usage{
marathon
}
\description{
Marathon times of male and female winners of the New York City Marathon
1970-1999. See \link{nyc_marathon} for a more updated dataset. We recommend not
using this dataset since the data source has been taken off the web.
}
\examples{
library(ggplot2)

ggplot(marathon, aes(x = time)) +
  geom_histogram(binwidth = 0.15)

ggplot(marathon, aes(y = time, x = gender)) +
  geom_boxplot()
}
\keyword{datasets}
