% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mlb_players_18.R
\docType{data}
\name{mlb_players_18}
\alias{mlb_players_18}
\title{Batter Statistics for 2018 Major League Baseball (MLB) Season}
\format{A data frame with 1270 observations on the following 19 variables.
\describe{ \item{list("name")}{Player name} \item{list("team")}{Team
abbreviation} \item{list("position")}{Position abbreviation: \code{1B} =
first base, \code{2B} = second base, \code{3B} = third base, \code{C} =
catcher, \code{CF} = center field (outfield), \code{DH} = designated hitter,
\code{LF} = left field (outfield), \code{P} = pitcher, \code{RF} = right
field (outfield), \code{SS} = shortstop.} \item{list("games")}{Number of
games played.} \item{list("AB")}{At bats.} \item{list("R")}{Runs.}
\item{list("H")}{Hits.} \item{list("doubles")}{Doubles.}
\item{list("triples")}{Triples.} \item{list("HR")}{Home runs.}
\item{list("RBI")}{Runs batted in.} \item{list("walks")}{Walks.}
\item{list("strike_outs")}{Strike outs.} \item{list("stolen_bases")}{Stolen
bases.} \item{list("caught_stealing_base")}{Number of times caught stealing
a base.} \item{list("AVG")}{Batting average.} \item{list("OBP")}{On-base
percentage.} \item{list("SLG")}{Slugging percentage.}
\item{list("OPS")}{On-base percentage plus slugging percentage.} }}
\source{
\url{http://mlb.mlb.com/stats}
}
\usage{
mlb_players_18
}
\description{
Batter statistics for 2018 Major League Baseball season.
}
\examples{

d <- subset(mlb_players_18, !position \%in\% c("P", "DH") & AB >= 100)
dim(d)

# _____ Per Position, No Further Grouping _____ #
plot(d$OBP ~ as.factor(d$position))
model <- lm(OBP ~ as.factor(position), d)
summary(model)
anova(model)

# _____ Simplified Analysis, Fewer Positions _____ #
pos <- list(c("LF", "CF", "RF"),
    c("1B", "2B", "3B", "SS"),
    "C")
POS <- c("OF", "IF", "C")
table(d$position)

# _____ On-Base Percentage Across Positions _____ #
out <- c()
gp  <- c()
for(i in 1:length(pos)){
  these <- which(d$position \%in\% pos[[i]])
  out   <- c(out, d$OBP[these])
  gp    <- c(gp, rep(POS[i], length(these)))
}
plot(out ~ as.factor(gp))
summary(lm(out ~ as.factor(gp)))
anova(lm(out ~ as.factor(gp)))

}
\seealso{
\code{\link{mlbbat10}}, \code{\link{mlb}}
}
\keyword{datasets}
