% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-nhanes.samp.adult.R
\docType{data}
\name{nhanes.samp.adult}
\alias{nhanes.samp.adult}
\title{Selection of participants 21 years of age or older from nhanes.samp.}
\format{
A dataframe with 135 rows and 76 variables
}
\source{
\url{https://CRAN.R-project.org/package=NHANES}.
}
\usage{
nhanes.samp.adult
}
\description{
The dataset NHANES (US National Health and Nutrition Examination Study) is
part of the CRAN package \code{NHANES} (author Randall Purim, rpruim@calvin.edu)
and contains 76 variables on 100,000 participants from surveys conducted
between 2009 and 2012. The surveys are part of a series conducted by the US
National Center for Health Statistics (NCHS) since the 1960's. See the \code{NHANES}
package documentation for more information about the surveys.
}
\details{
The dataset NHANES is a weighted sample from the full survey dataset constructed
so that it may be treated as a random sample of the US population. The dataset
nhanes.samp.adult contains data from the 135 participants 21 years of age or
older from the nhanes.samp dataset. See the \code{NHANES} package for variable
definitions and coding.
}
\references{
Pruim R (2015). \emph{NHANES: Data from the US National Health and
Nutrition Examination Study}. R package version 2.1.0,
}
\keyword{datasets}
