% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-nhanes.samp.R
\docType{data}
\name{nhanes.samp}
\alias{nhanes.samp}
\title{Random sample of 200 observations from the dataset NHANES.}
\format{
A dataframe with 200 rows and 76 variables
}
\source{
\url{https://CRAN.R-project.org/package=NHANES}.
}
\usage{
nhanes.samp
}
\description{
The dataset NHANES (US National Health and Nutrition Examination Study) is
part of the CRAN package \code{NHANES} (author Randall Purim, rpruim@calvin.edu) and
contains 76 variables on 100,000 participants from surveys conducted between
2009 and 2012. The surveys are part of a series conducted by the US National
Center for Health Statistics (NCHS) since the 1960's. See the \code{NHANES} package
documentation for more information about the surveys.
}
\details{
The dataset NHANES is a weighted sample from the full survey dataset constructed
so that it may be treated as a random sample of the US population. The dataset
nhanes.samp contains data from a random sample of size 200 from NHANES and
all 76 variables. See the NHANES package for variable definitions and coding.
}
\references{
Pruim R (2015). \emph{NHANES: Data from the US National Health and
Nutrition Examination Study}. R package version 2.1.0,
}
\keyword{datasets}
