% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-nuclear_survey.R
\docType{data}
\name{nuclear_survey}
\alias{nuclear_survey}
\title{Nuclear Arms Reduction Survey}
\format{
A data frame with 1028 observations on the following variable.
\describe{\item{arms_reduction}{Responses of \code{favor} or
\code{against}.} }
}
\source{
Gallup report: In U.S., 56 percent Favor U.S.-Russian Nuclear Arms Reductions.
Available at \url{https://news.gallup.com/poll/161198/favor-russian-nuclear-arms-reductions.aspx}.
}
\usage{
nuclear_survey
}
\description{
A simple random sample of 1,028 US adults in March 2013 found that 56\\%
support nuclear arms reduction.
}
\examples{

table(nuclear_survey)
}
\keyword{datasets}
