% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-offshore_drilling.R
\docType{data}
\name{offshore_drilling}
\alias{offshore_drilling}
\alias{offshore.drilling}
\title{California poll on drilling off the California coast}
\format{
A data frame with 827 observations on the following 2 variables.
\describe{\item{position}{a factor with levels \code{do not know}
\code{oppose} \code{support}} \item{college_grad}{a factor with
levels \code{no} \code{yes}} }
}
\source{
Survey USA, Election Poll #16804, data collected July 8-11, 2010.
}
\usage{
offshore_drilling
}
\description{
A 2010 survey asking a randomly sample of registered voters in California
for their position on drilling for oil and natural gas off the Coast of
California.
}
\examples{

offshore_drilling
}
\keyword{datasets}
