% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-openintro-colors.R
\docType{data}
\name{openintro_colors}
\alias{openintro_colors}
\title{OpenIntro colors}
\format{
A named character string with 9 elements: "blue", "green", "pink", "yellow", "red", "black", "gray", "lgray
}
\usage{
openintro_colors
}
\description{
A character string of full colors from \code{IMSCOL[,1]}
}
\examples{

openintro_colors

openintro_colors["blue"]
}
\keyword{datasets}
