% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-openintro.R
\name{openintro_cols}
\alias{openintro_cols}
\title{Function to extract OpenIntro IMS colors as hex codes}
\usage{
openintro_cols(...)
}
\arguments{
\item{...}{Character names of \link{openintro_colors}}
}
\description{
Uses full colors from \link{IMSCOL}
}
\examples{

openintro_cols("blue")
openintro_cols("red")
}
