% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-openintro.R
\name{openintro_pal}
\alias{openintro_pal}
\title{Return function to interpolate an OpenIntro IMS color palette}
\usage{
openintro_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in \link{openintro_palettes}}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to \code{\link[grDevices:colorRamp]{grDevices::colorRampPalette()}}}
}
\description{
Not exported
}
