% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-openintro-palettes.R
\docType{data}
\name{openintro_palettes}
\alias{openintro_palettes}
\title{OpenIntro palettes}
\format{
A list with 8 color palettes: main, two, three, four, five, cool, hot, gray
}
\usage{
openintro_palettes
}
\description{
A list with  OpenIntro color palettes
}
\examples{

openintro_palettes

openintro_palettes$main
openintro_palettes$three
openintro_palettes$cool
openintro_palettes$hot
}
\keyword{datasets}
