% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-oscars.R
\docType{data}
\name{oscars}
\alias{oscars}
\title{Oscar winners, 1929 to 2018}
\format{A tibble with 182 observations on the following 10 variables.
\describe{\item{gender}{Gender of winner, \code{female} or
\code{male}.} \item{oscar_no}{Denotes which Oscar ceremony.}
\item{oscar_yr}{Denotes which Oscar year.} \item{name}{Name
of winning actor or actress.} \item{movie}{Name of movie actor or
actress got the Oscar for.} \item{age}{Age at which the actor or
actress won the Oscar.} \item{birth_pl}{State where the actor or
actress was born, country if foreign.} \item{birth_mo}{Birth month
of actor or actress.} \item{birth_d}{Birth day of actor or actress.}
\item{birth_y}{Birth year of actor or actress.} }}
\source{
Journal of Statistical Education,
http://www.amstat.org/publications/jse/datasets/oscar2009.dat.txt (no longer
valid), updated through 2012 using information from Wikipedia.org.
}
\usage{
oscars
}
\description{
Best actor and actress Oscar winners from 1929 to 2018
}
\details{
Although there have been only 84 Oscar ceremonies until 2012, there are 85
male winners and 85 female winners because ties happened on two occasions
(1933 for the best actor and 1969 for the best actress).
}
\examples{

library(ggplot2)
library(dplyr)

ggplot(oscars, aes(x = gender, y = age)) +
  geom_boxplot()

ggplot(oscars, aes(x = factor(birth_mo))) +
  geom_bar()

oscars \%>\%
  count(birth_pl, sort = TRUE)

}
\keyword{datasets}
