\name{outliers}
\alias{outliers}
\docType{data}
\title{Simulated data sets for different types of outliers}
\description{Data sets for showing different types of outliers}
\usage{data("outliers")}
\format{
  A data frame with 50 observations on the following 5 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
    \item{\code{xInf}}{a numeric vector}
    \item{\code{yLev}}{a numeric vector}
    \item{\code{yOut}}{a numeric vector}
  }
}
\examples{
data(outliers)
## maybe str(outliers) ; plot(outliers) ...
}
\keyword{datasets}