% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageData2Csv.R
\name{PackageData2Csv}
\alias{PackageData2Csv}
\title{Create a CSV variant of .rda files}
\usage{
PackageData2Csv(pkg, dir = "data-csv", overwrite = FALSE)
}
\arguments{
\item{pkg}{The R package where we'd like to generate CSVs of any
data frames.}

\item{dir}{A character string representing the path to the folder.
where the CSV files should be written. If no such directory
exists, one will be created (recursively).}

\item{overwrite}{Boolean to indicate if to overwrite any existing
files that have conflicting names in the directory specified.}
}
\description{
The function should be run with a path to a package directory.
It will then look through the \code{data} directory of the package,
and for all data sets that are data frames, create CSV variants
in a \code{data-csv} directory.
}
\examples{

\dontrun{
  PackageData2Csv("openintro")
  list.files("data-csv")
}
}
