% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-penny_ages.R
\docType{data}
\name{penny_ages}
\alias{penny_ages}
\alias{penny.ages}
\title{Penny Ages}
\format{
A data frame with 648 observations on the following 2 variables.
\describe{
\item{year}{Penny's year.}
\item{age}{Age as of 2004.}
}
}
\usage{
penny_ages
}
\description{
Sample of pennies and their ages. Taken in 2004.
}
\examples{

hist(penny_ages$year)
}
\keyword{datasets}
