% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotWLine.R
\name{PlotWLine}
\alias{PlotWLine}
\title{Plot data and add a regression line}
\usage{
PlotWLine(
  x,
  y,
  xlab = "",
  ylab = "",
  col = fadeColor(4, "88"),
  cex = 1.2,
  pch = 20,
  n = 4,
  nMax = 4,
  yR = 0.1,
  axes = TRUE,
  ...
)
}
\arguments{
\item{x}{Predictor variable.}

\item{y}{Outcome variable.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{col}{Color of points.}

\item{cex}{Size of points.}

\item{pch}{Plotting character.}

\item{n}{The preferred number of axis labels.}

\item{nMax}{The maximum number of axis labels.}

\item{yR}{y-limit buffer factor.}

\item{axes}{Boolean to indicate whether or not to include axes.}

\item{\dots}{Passed to \code{plot}.}
}
\description{
Plot data and add a regression line.
}
\examples{

PlotWLine(1:10, seq(-5, -2, length.out = 10) + rnorm(10))

}
\seealso{
\code{\link{makeTube}}
}
