\name{president}
\alias{president}
\docType{data}
\title{
United States Presidental History
}
\description{
Summary of the changes in the president and vice president for the United States of America.
}
\usage{data(president)}
\format{
  A data frame with 67 observations on the following 5 variables.
  \describe{
    \item{\code{potus}}{President of the United States}
    \item{\code{party}}{Political party of the president}
    \item{\code{start}}{Start year}
    \item{\code{end}}{End year}
    \item{\code{vpotus}}{Vice President of the United States}
  }
}
\source{
Presidents of the United States (table) -- infoplease.com (visited: Nov 2nd, 2010)

\url{http://www.infoplease.com/ce6/history/A0840075.html}
}
\examples{
data(president)
}
\keyword{datasets}
