% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prison.R
\docType{data}
\name{prison}
\alias{prison}
\title{Prison isolation experiment}
\format{
A data frame with 14 observations on the following 6 variables.
\describe{\item{PreTrt1}{Pre-treatment 1.} \item{PostTrt1}{Post-treatment
1.} \item{PreTrt2}{Pre-treatment 2.} \item{PostTrt2}{Post-treatment 2.}
\item{PreTrt3}{Pre-treatment 3.} \item{PostTrt3}{Post-treatment 3.} }
}
\source{
\url{http://stat.duke.edu/resources/datasets/prison-isolation}
}
\usage{
prison
}
\description{
Subjects from Central Prison in Raleigh, NC, volunteered for an experiment
involving an "isolation" experience. The goal of the experiment was to find
a treatment that reduces subjects' psychopathic deviant T scores. This score
measures a person's need for control or their rebellion against control, and
it is part of a commonly used mental health test called the Minnesota
Multiphasic Personality Inventory (MMPI) test.
}
\examples{


prison


}
\keyword{datasets}
