\name{prof.evals}
\alias{prof.evals}
\alias{prof_evals}
\docType{data}
\title{
  Professor evaluations and beauty
}
\description{
  Data set from a paper on professor evaluations and beauty scores.
}
\usage{data("prof.evals")}
\format{
  A data frame with 463 observations on the following 64 variables.
  \describe{
    \item{\code{tenured}}{Tenured indicator.}
    \item{\code{profnumber}}{Professor number.}
    \item{\code{minority}}{Minority.}
    \item{\code{age}}{Age.}
    \item{\code{beautyf2upper}}{A numeric vector.}
    \item{\code{beautyflowerdiv}}{A numeric vector.}
    \item{\code{beautyfupperdiv}}{A numeric vector.}
    \item{\code{beautym2upper}}{A numeric vector.}
    \item{\code{beautymlowerdiv}}{A numeric vector.}
    \item{\code{beautymupperdiv}}{A numeric vector.}
    \item{\code{btystdave}}{A numeric vector.}
    \item{\code{btystdf2u}}{A numeric vector.}
    \item{\code{btystdfl}}{A numeric vector.}
    \item{\code{btystdfu}}{A numeric vector.}
    \item{\code{btystdm2u}}{A numeric vector.}
    \item{\code{btystdml}}{A numeric vector.}
    \item{\code{btystdmu}}{A numeric vector.}
    \item{\code{class1}}{Class 1.}
    \item{\code{class2}}{Class 2.}
    \item{\code{class3}}{Class 3.}
    \item{\code{class4}}{Class 4.}
    \item{\code{class5}}{Class 5.}
    \item{\code{class6}}{Class 6.}
    \item{\code{class7}}{Class 7.}
    \item{\code{class8}}{Class 8.}
    \item{\code{class9}}{Class 9.}
    \item{\code{class10}}{Class 10.}
    \item{\code{class11}}{Class 11.}
    \item{\code{class12}}{Class 12.}
    \item{\code{class13}}{Class 13.}
    \item{\code{class14}}{Class 14.}
    \item{\code{class15}}{Class 15.}
    \item{\code{class16}}{Class 16.}
    \item{\code{class17}}{Class 17.}
    \item{\code{class18}}{Class 18.}
    \item{\code{class19}}{Class 19.}
    \item{\code{class20}}{Class 20.}
    \item{\code{class21}}{Class 21.}
    \item{\code{class22}}{Class 22.}
    \item{\code{class23}}{Class 23.}
    \item{\code{class24}}{Class 24.}
    \item{\code{class25}}{Class 25.}
    \item{\code{class26}}{Class 26.}
    \item{\code{class27}}{Class 27.}
    \item{\code{class28}}{Class 28.}
    \item{\code{class29}}{Class 29.}
    \item{\code{class30}}{Class 30.}
    \item{\code{courseevaluation}}{Course evaluation.}
    \item{\code{didevaluation}}{Did evaluation.}
    \item{\code{female}}{Female indicator.}
    \item{\code{formal}}{Formal.}
    \item{\code{fulldept}}{Full department.}
    \item{\code{lower}}{Lower.}
    \item{\code{multipleclass}}{Multiple class.}
    \item{\code{nonenglish}}{Non-English.}
    \item{\code{onecredit}}{One credit.}
    \item{\code{percentevaluating}}{Percent evaluating.}
    \item{\code{profevaluation}}{Professor evaluation.}
    \item{\code{students}}{Students}
    \item{\code{tenuretrack}}{Tenure-track indicator.}
    \item{\code{blkandwhite}}{Black and white.}
    \item{\code{btystdvariance}}{Beauty standard variance.}
    \item{\code{btystdavepos}}{Beauty standard average position.}
    \item{\code{btystdaveneg}}{Beauty standard average negative.}
  }
}
\source{
  Hamermesh DS, Parker A. 2005. "Beauty in the classroom:
  Instructors pulchritude and putative pedagogical productivity".
  Economics of Education Review 24.4:369-376. 
}
\examples{
data(prof.evals)
}
\seealso{
See also \code{\link{evals}} for simplified version of dataset.
}
\keyword{datasets}
