% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prof_evals.R
\docType{data}
\name{prof_evals}
\alias{prof_evals}
\alias{prof.evals}
\title{Professor evaluations and beauty}
\format{
A data frame with 463 observations on the following 64 variables.
\describe{\item{tenured}{Tenured indicator.} \item{profnumber}{Professor
number.} \item{minority}{Minority.} \item{age}{Age.} \item{beautyf2upper}{A
numeric vector.} \item{beautyflowerdiv}{A numeric vector.}
\item{beautyfupperdiv}{A numeric vector.} \item{beautym2upper}{A numeric
vector.} \item{beautymlowerdiv}{A numeric vector.} \item{beautymupperdiv}{A
numeric vector.} \item{btystdave}{A numeric vector.} \item{btystdf2u}{A
numeric vector.} \item{btystdfl}{A numeric vector.} \item{btystdfu}{A
numeric vector.} \item{btystdm2u}{A numeric vector.} \item{btystdml}{A
numeric vector.} \item{btystdmu}{A numeric vector.} \item{class1}{Class 1.}
\item{class2}{Class 2.} \item{class3}{Class 3.} \item{class4}{Class 4.}
\item{class5}{Class 5.} \item{class6}{Class 6.} \item{class7}{Class 7.}
\item{class8}{Class 8.} \item{class9}{Class 9.} \item{class10}{Class 10.}
\item{class11}{Class 11.} \item{class12}{Class 12.} \item{class13}{Class
13.} \item{class14}{Class 14.} \item{class15}{Class 15.}
\item{class16}{Class 16.} \item{class17}{Class 17.} \item{class18}{Class
18.} \item{class19}{Class 19.} \item{class20}{Class 20.}
\item{class21}{Class 21.} \item{class22}{Class 22.} \item{class23}{Class
23.} \item{class24}{Class 24.} \item{class25}{Class 25.}
\item{class26}{Class 26.} \item{class27}{Class 27.} \item{class28}{Class
28.} \item{class29}{Class 29.} \item{class30}{Class 30.}
\item{courseevaluation}{Course evaluation.} \item{didevaluation}{Did
evaluation.} \item{female}{Female indicator.} \item{formal}{Formal.}
\item{fulldept}{Full department.} \item{lower}{Lower.}
\item{multipleclass}{Multiple class.} \item{nonenglish}{Non-English.}
\item{onecredit}{One credit.} \item{percentevaluating}{Percent evaluating.}
\item{profevaluation}{Professor evaluation.} \item{students}{Students}
\item{tenuretrack}{Tenure-track indicator.} \item{blkandwhite}{Black and
white.} \item{btystdvariance}{Beauty standard variance.}
\item{btystdavepos}{Beauty standard average position.}
\item{btystdaveneg}{Beauty standard average negative.} }
}
\source{
Hamermesh DS, Parker A. 2005. "Beauty in the classroom: Instructors
pulchritude and putative pedagogical productivity". Economics of Education
Review 24.4:369-376.
}
\usage{
prof_evals
}
\description{
Data set from a paper on professor evaluations and beauty scores.
}
\examples{


prof_evals


}
\seealso{
See also \code{\link{evals}} for simplified version of dataset.
}
\keyword{datasets}
