% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-race-justice.R
\docType{data}
\name{race_justice}
\alias{race_justice}
\title{Yahoo! News Race and Justice poll results.}
\format{
A data frame with 4 rows and 5 variables.
\describe{
\item{race}{Race of respondent.}
\item{yes}{Number of respondents who answered "yes".}
\item{no}{Number of respondents who answered "no".}
\item{not_sure}{Number of respondents who answered "not sure".}
\item{unweighted_n}{Sample size for each race. }
}
}
\source{
\href{https://docs.cdn.yougov.com/s23agrrx47/20200531_yahoo_race_and_justice_crosstabs.pdf}{Yahoo! News Race and Justice - May 31, 2020}
}
\usage{
race_justice
}
\description{
Results from a Yahoo! News poll conducted by YouGov on May 29-39, 2020.
In total 1060 U.S. adults were asked a series of questions regarding race and justice in the wake of
the killing of George Floyd by a police officer. Results in this data set are percentages for the question,
"Do you think Blacks and Whites receive equal treatment from the police?" For this particular question
there were 1059 respondents.
}
\examples{

barplot(race_justice$yes/race_justice$unweighted_n,
     main = "Do police treat Black citizens and White citizens equally?",
     xlab = "Race",
     ylab = "Yes (percent)",
     names.arg = c("White","Black","Hispanic","Other"))

barplot(race_justice$no/race_justice$unweighted_n,
     main = "Do police treat Black citizens and White citizens equally?",
     xlab = "Race",
     ylab = "No (percent)",
     names.arg = c("White","Black","Hispanic","Other"))
}
\keyword{datasets}
