\name{res.demo.1}
\alias{res.demo.1}
\docType{data}
\title{Simulated data for regression}
\description{Simulated data for regression.}
\usage{data("res.demo.1")}
\format{
  A data frame with 100 observations on the following 3 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{yLin}}{a numeric vector}
    \item{\code{yFanBack}}{a numeric vector}
  }
}
\examples{
data(res.demo.1)
## maybe str(res.demo.1) ; plot(res.demo.1) ...
}
\keyword{datasets}