\name{res.demo.2}
\alias{res.demo.2}
\docType{data}
\title{Simulated data for regression}
\description{Simulated data for regression.}
\usage{data("res.demo.2")}
\format{
  A data frame with 300 observations on the following 3 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{yFan}}{a numeric vector}
    \item{\code{yLog}}{a numeric vector}
  }
}
\examples{
data(res.demo.2)
## maybe str(res.demo.2) ; plot(res.demo.2) ...
}
\keyword{datasets}