% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-res_demo.R
\docType{data}
\name{res_demo_1}
\alias{res_demo_1}
\title{Simulated data for regression}
\format{A tibble with 100 observations on the following 3 variables.
\describe{
  \item{x}{a numeric vector}
  \item{yLin}{a numeric vector}
  \item{yFanBack}{a numeric vector}
  }}
\usage{
res_demo_1
}
\description{
Simulated data for regression.
}
\examples{

res_demo_1

}
\keyword{datasets}
