% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-res_demo.R
\docType{data}
\name{res_demo_2}
\alias{res_demo_2}
\title{Simulated data for regression}
\format{
A data frame with 300 observations on the following 3 variables.
\describe{
\item{x}{a numeric vector}
\item{y_fan}{a numeric vector}
\item{y_log}{a numeric vector}
}
}
\usage{
res_demo_2
}
\description{
Simulated data for regression
}
\examples{

res_demo_2

}
\keyword{datasets}
