\name{sat.improve}
\alias{sat.improve}
\docType{data}
\title{Simulated data for SAT score improvement}
\description{Fake data for score improvements from students who took a course from an SAT score improvement company.}
\usage{data("sat.improve")}
\format{
  A data frame with 30 observations on the following variable.
  \describe{
    \item{\code{sat_improve}}{a numeric vector}
  }
}
\examples{
data(sat.improve)
## maybe str(sat.improve) ; plot(sat.improve) ...
}
\keyword{datasets}