% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sat_improve.R
\docType{data}
\name{sat_improve}
\alias{sat_improve}
\title{Simulated data for SAT score improvement}
\format{
A data frame with 30 observations on the following variable.
\describe{
  \item{sat_improve}{a numeric vector}
  }
}
\usage{
sat_improve
}
\description{
Fake data for score improvements from students who took a course from an SAT
score improvement company.
}
\examples{

sat_improve

}
\keyword{datasets}
