% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sibling_heights.R
\docType{data}
\name{sibling_heights}
\alias{sibling_heights}
\title{Sibling heights}
\format{
A data frame with 11 rows and 3 variables.
\describe{
  \item{family}{An index variable to distinguish families in the study.}
  \item{brother}{Height of the male sibling in inches.}
  \item{sister}{Height of the female sibling in inches.}
}
}
\source{
Pearson, K. and Lee, A. (1902-3) On the laws of inheritance in man. Biometrika, 2, 357.
}
\usage{
sibling_heights
}
\description{
Data collected from 11 families on the heights of male and female siblings
collected by Pearson and Lee to study correlation between statures of siblings.
}
\examples{

library(ggplot2)

ggplot(data = sibling_heights, aes(x = brother, y = sister))+
  geom_point()+
  geom_smooth(method = "lm", se = FALSE)+
  labs(
    title = "Sibling heights",
    x = "Height of brother (in.)",
    y = "Height of sister (in.)"
    )

}
\keyword{datasets}
