% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sleep_deprivation.R
\docType{data}
\name{sleep_deprivation}
\alias{sleep_deprivation}
\title{Survey on sleep deprivation and transportation workers}
\format{
A data frame with 1087 observations on the following 2 variables.
\describe{
  \item{sleep}{a factor with levels \code{<6}, \code{6-8}, and \code{>8}}
  \item{profession}{a factor with levels \code{bus / taxi / limo drivers},
  \code{control}, \code{pilots}, \code{train operators}, \code{truck drivers}}
  }
}
\source{
National Sleep Foundation, 2012 Sleep in America Poll:
Transportation Workers' Sleep, 2012.
\url{https://sleepfoundation.org/sleep-polls-data/sleep-in-america-poll/2012-transportation-workers-and-sleep}
}
\usage{
sleep_deprivation
}
\description{
The National Sleep Foundation conducted a survey on the sleep habits of
randomly sampled transportation workers and a control sample of
non-transportation workers.
}
\examples{

sleep_deprivation

}
\keyword{datasets}
