% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-soda.R
\docType{data}
\name{soda}
\alias{soda}
\title{soda}
\format{
A data frame with 60 observations on the following 2 variables.
\describe{
\item{drink}{Soda preference, cola or orange.}
\item{location}{Is the person from the West coast or East coast?}
}
}
\usage{
soda
}
\description{
A randomly generated dataset of soda preference (cola or orange) based on location.
}
\examples{

library(dplyr)

soda |>
  count(location, drink)
}
\keyword{datasets}
