% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sp500_seq.R
\docType{data}
\name{sp500_seq}
\alias{sp500_seq}
\title{S&P 500 stock data}
\format{
A data frame with 2948 observations on the following variable.
\describe{
  \item{race}{a factor with levels \code{1}, \code{2}, \code{3}, \code{4},
  \code{5}, \code{6}, and \code{7+}}
  }
}
\source{
\url{http://www.google.com/finance}
}
\usage{
sp500_seq
}
\description{
Daily stock returns from the S&P500 for 1990-2011 can be used to assess
whether stock activity each day is independent of the stock's behavior on
previous days. We label each day as Up or Down (D) depending on whether the
market was up or down that day. For example, consider the following changes
in price, their new labels of up and down, and then the number of days that
must be observed before each Up day.
}
\examples{

sp500_seq

}
\keyword{datasets}
