% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-speed_gender_height.R
\docType{data}
\name{speed_gender_height}
\alias{speed_gender_height}
\title{Speed, gender, and height of 1325 students}
\format{
A data frame with 1325 observations on the following 3 variables.
\describe{
\item{speed}{a numeric vector}
\item{gender}{a factor with levels \code{female} and \code{male}}
\item{height}{a numeric vector}
}
}
\usage{
speed_gender_height
}
\description{
1,325 UCLA students were asked to fill out a survey where they were asked
about their height, fastest speed they have ever driven, and gender.
}
\examples{

speed_gender_height

}
\keyword{datasets}
