% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state2abbr.R
\name{state2abbr}
\alias{state2abbr}
\title{Convert state names to abbreviations}
\usage{
state2abbr(state)
}
\arguments{
\item{state}{A vector of state name, where there is a little fuzzy matching.}
}
\value{
Returns a vector of the same length with the corresponding state
names or abbreviations.
}
\description{
Two utility functions. One converts state names to the state abbreviations,
and the second does the opposite.
}
\examples{

state2abbr("Minnesota")

# Some spelling/capitalization errors okay
state2abbr("mINnesta")

}
\seealso{
\code{\link{abbr2state}}, \code{\link{county}}, \code{\link{county_complete}}
}
\author{
David Diez
}
\keyword{Abbreviation}
\keyword{State}
