\name{stats.scores}
\alias{stats.scores}
\docType{data}
\title{Final exam scores for twenty students}
\description{Scores range from 57 to 94.}
\usage{data("stats.scores")}
\format{
  A data frame with 20 observations on the following variable.
  \describe{
    \item{\code{scores}}{a numeric vector}
  }
}
\examples{
data(stats.scores)
## maybe str(stats.scores) ; plot(stats.scores) ...
}
\keyword{datasets}