% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-stats_scores.R
\docType{data}
\name{stats_scores}
\alias{stats_scores}
\title{Final exam scores for twenty students}
\format{
A data frame with 20 observations on the following variable.
\describe{
\item{scores}{a numeric vector}
}
}
\usage{
stats_scores
}
\description{
Scores range from 57 to 94.
}
\examples{

stats_scores
}
\keyword{datasets}
