% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sugar.levels.A.R
\docType{data}
\name{sugar.levels.A}
\alias{sugar.levels.A}
\title{Simulated fasting blood sugar levels for 100 residents of a hypothetical neighborhood labeled A.}
\format{
A tibble with 100 rows and 1 variable
\describe{
\item{\code{fasting.blood.sugar}}{Numeric, simulated fasting blood sugar in nmol/L}
}
}
\usage{
sugar.levels.A
}
\description{
Simulated individual fasting blood sugar levels (nmol/L) drawn from  a normal
distribution. Generally, normal fasting blood sugar level are between 3.0 - 5.6 nmol/L;
levels in the range 5.6 - 6.9 nmol/L are considered pre-diabetes. These data
and sugar.levels.B are used in the Unit 4 lab of Introductory Statistics for
the Life and Biomedical Sciences (ISLBS). See \url{https://github.com/OI-Biostat/oi_biostat_labs} for the full set of labs.
}
\keyword{datasets}
