% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-supreme_court.R
\docType{data}
\name{supreme_court}
\alias{supreme_court}
\title{Supreme Court approval rating}
\format{
A data frame with 976 observations on the following variable.
\describe{
\item{answer}{a factor with levels \code{approve} and \code{not}}
}
}
\source{
\url{https://www.nytimes.com/2012/06/08/us/politics/44-percent-of-americans-approve-of-supreme-court-in-new-poll.html}
}
\usage{
supreme_court
}
\description{
Summary of a random survey of 976 people.
}
\examples{

supreme_court

}
\keyword{datasets}
