% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-thanksgiving_spend.R
\docType{data}
\name{thanksgiving_spend}
\alias{thanksgiving_spend}
\alias{thanksgiving.spend}
\alias{tgSpending}
\title{Thanksgiving spending, simulated based on Gallup poll.}
\format{
A data frame with 436 observations on the following 1 variable.
\describe{
\item{spending}{Amount of spending, in US dollars.}
}
}
\usage{
thanksgiving_spend
}
\description{
This entry gives simulated spending data for Americans during Thanksgiving
in 2009 based on findings of a Gallup poll.
}
\examples{

library(ggplot2)

ggplot(thanksgiving_spend, aes(x = spending)) +
  geom_histogram(binwidth = 20)

}
\keyword{datasets}
\keyword{distribution}
\keyword{histogram}
