\name{toohey}
\alias{toohey}
\docType{data}
\title{Simulated polling data set}
\description{Simulated data for a fake political candidate.}
\usage{data("toohey")}
\format{
  A data frame with 500 observations on the following variable.
  \describe{
    \item{\code{voteFor}}{a factor with levels \code{no} \code{yes}}
  }
}
\examples{
data(toohey)
## maybe str(toohey) ; plot(toohey) ...
}
\keyword{datasets}