% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-transplant.R
\docType{data}
\name{transplant}
\alias{transplant}
\title{Transplant consultant success rate (fake data)}
\format{
A data frame with 62 observations on the following variable.
\describe{
  \item{outcome}{a factor with levels \code{complications} \code{okay}}
  }
}
\usage{
transplant
}
\description{
Summarizing whether there was or was not a complication for 62 patients who
used a particular medical consultant.
}
\examples{

transplant

}
\keyword{datasets}
