\name{trees}
\alias{trees}
\docType{data}
\title{Summary of 32 trees}
\description{Variables include girth, height, and volume.}
\usage{data("trees")}
\format{
  A data frame with 31 observations on the following 3 variables.
  \describe{
    \item{\code{Girth}}{Diameter, in inches}
    \item{\code{Height}}{Height, in feet.}
    \item{\code{Volume}}{Volume, in cubic feet.}
  }
}
\source{Source: R Dataset. http://stat.ethz.ch/R-manual/R-patched/library/datasets/html/trees.html}
\examples{
data(trees)
## maybe str(trees) ; plot(trees) ...
}
\keyword{datasets}