% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-unemploy_pres.R
\docType{data}
\name{unemploy_pres}
\alias{unemploy_pres}
\title{President's party performance and unemployment rate}
\format{
A data frame with 29 observations on the following 5 variables.
\describe{
\item{year}{a numeric vector}
\item{potus}{The president in office.}
\item{party}{President's party.}
\item{unemp}{Unemployment rate.}
\item{change}{Change in House seats for the president's party.}
}
}
\source{
Wikipedia.
}
\usage{
unemploy_pres
}
\description{
Covers midterm elections.
}
\examples{

unemploy_pres

}
\keyword{datasets}
