% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-urban_rural_pop.R
\docType{data}
\name{urban_rural_pop}
\alias{urban_rural_pop}
\title{State summary info}
\format{A tibble with 51 observations on the following 5 variables.
\describe{
  \item{state}{US state.}
  \item{urban_in}{a numeric vector}
  \item{urban_out}{a numeric vector}
  \item{rural_farm}{a numeric vector}
  \item{rural_nonfarm}{a numeric vector}
  }}
\source{
US census.
}
\usage{
urban_rural_pop
}
\description{
Census info for the 50 US states plus DC.
}
\examples{

urban_rural_pop

}
\keyword{datasets}
