% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-yawn.R
\docType{data}
\name{yawn}
\alias{yawn}
\title{Contagiousness of yawning}
\format{
A data frame with 50 observations on the following 2 variables.
\describe{
\item{result}{a factor with levels \code{not yawn} \code{yawn}}
\item{group}{a factor with levels \code{ctrl} \code{trmt}}
}
}
\source{
MythBusters, Season 3, Episode 28.
}
\usage{
yawn
}
\description{
An experiment conducted by the MythBusters, a science entertainment TV
program on the Discovery Channel, tested if a person can be subconsciously
influenced into yawning if another person near them yawns. 50 people were
randomly assigned to two groups: 34 to a group where a person near them
yawned (treatment) and 16 to a group where there wasn't a person yawning
near them (control).
}
\examples{

yawn

}
\keyword{datasets}
