% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbr2state.R
\name{abbr2state}
\alias{abbr2state}
\alias{state2abbr}
\title{Convert state names to abbreviations and back again}
\usage{
abbr2state(abbr)
}
\arguments{
\item{abbr}{A vector of state abbreviation.}

\item{state}{A vector of state name, where there is a little fuzzy matching.}
}
\value{
Returns a vector of the same length with the corresponding state
names or abbreviations.
}
\description{
Two utility functions. One converts state names to the state abbreviations,
and the second does the opposite.
}
\examples{

state2abbr("Minnesota")
abbr2state("MN")

#_____ Some Spelling/Capitalization Errors Okay _____#
state2abbr("mINnesta")

}
\seealso{
\code{\link{county}}, \code{\link{county_complete}}
}
\author{
David Diez
}
\keyword{Abbreviation}
\keyword{State}
