% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-books.R
\docType{data}
\name{books}
\alias{books}
\title{Sample of books on a shelf}
\format{
A data frame with 95 observations on the following 2 variables.
\describe{
  \item{type}{a factor with levels \code{fiction} and \code{nonfiction}}
  \item{format}{a factor with levels \code{hardcover} and \code{paperback}}
  }
}
\usage{
books
}
\description{
Simulated data set.
}
\examples{

table(books)

}
\keyword{datasets}
