% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-burger.R
\docType{data}
\name{burger}
\alias{burger}
\title{Burger preferences}
\format{
A data frame with 500 observations on the following 2 variables.
\describe{
\item{best_burger_place}{Burger place.}
\item{gender}{a factor with levels \code{Female} and \code{Male}}
}
}
\source{
SurveyUSA, Results of SurveyUSA News Poll #17718, data collected on
December 2, 2010.
}
\usage{
burger
}
\description{
Sample burger place preferences versus gender.
}
\examples{

table(burger)
}
\keyword{datasets}
