% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cchousing.R
\docType{data}
\name{cchousing}
\alias{cchousing}
\title{Community college housing (simulated data)}
\format{
A data frame with 75 observations on the following variable.
\describe{
\item{price}{Monthly housing price, simulated.}
}
}
\usage{
cchousing
}
\description{
These are simulated data and intended to represent housing prices of
students at a community college.
}
\examples{

hist(cchousing$price)
}
\keyword{datasets}
