% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-china.R
\docType{data}
\name{china}
\alias{china}
\title{Child care hours}
\format{
A data frame with 9788 observations on the following 3 variables.
\describe{
\item{gender}{a numeric vector} \item{edu}{a numeric vector}
\item{child_care}{a numeric vector}
}
}
\source{
UNC Carolina Population Center, China Health and Nutrition Survey,
2006.
}
\usage{
china
}
\description{
The China Health and Nutrition Survey aims to examine the effects of the
health, nutrition, and family planning policies and programs implemented by
national and local governments.
}
\examples{

summary(china)

}
\keyword{datasets}
