% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChiSquareTail.R
\name{ChiSquareTail}
\alias{ChiSquareTail}
\title{Plot upper tail in chi-square distribution}
\usage{
ChiSquareTail(
  U,
  df,
  xlim = c(0, 10),
  col = fadeColor("black", "22"),
  axes = TRUE,
  ...
)
}
\arguments{
\item{U}{Cut off for the upper tail.}

\item{df}{Degrees of freedom.}

\item{xlim}{Limits for the plot.}

\item{col}{Color of the shading.}

\item{axes}{Whether to plot an x-axis.}

\item{\dots}{Currently ignored.}
}
\value{
Nothing is returned from the function.
}
\description{
Plot a chi-square distribution and shade the upper tail.
}
\examples{

data(COL)
ChiSquareTail(11.7,
              7,
              c(0, 25),
              col = COL[1])

}
\seealso{
\code{\link{normTail}}
}
\author{
David Diez
}
