% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-classdata.R
\docType{data}
\name{classdata}
\alias{classdata}
\title{Simulated class data}
\format{
A data frame with 164 observations on the following 2 variables.
\describe{
\item{m1}{Represents a first midterm score.}
\item{lecture}{Three classes: \code{a}, \code{b}, and \code{c}.}
}
}
\usage{
classdata
}
\description{
This data is simulated and is meant to represent students scores from three
different lectures who were all given the same exam.
}
\examples{

anova(lm(m1 ~ lecture, classdata))

}
\references{
OpenIntro Statistics, Chapter 8.
}
\keyword{datasets}
