% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-country_iso.R
\docType{data}
\name{country_iso}
\alias{country_iso}
\title{Country ISO information}
\format{
A data frame with 249 observations on the following 4 variables.
\describe{
\item{country_code}{Two-letter ISO country code.}
\item{country_name}{Country name.}
\item{year}{Year the two-letter ISO country code was assigned.}
\item{top_level_domain}{ op-level domain name.}
}
}
\source{
Wikipedia, retrieved 2018-11-18.
\url{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}
}
\usage{
country_iso
}
\description{
Country International Organization for Standardization (ISO) information.
}
\examples{

country_iso
}
\keyword{datasets}
