% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cpr.R
\docType{data}
\name{cpr}
\alias{cpr}
\title{CPR dataset}
\format{
A data frame with 90 observations on the following 2 variables.
\describe{
\item{group}{a factor with levels \code{control} and \code{treatment}}
\item{outcome}{a factor with levels \code{died} and \code{survived}}
}
}
\source{
Efficacy and safety of thrombolytic therapy after initially
unsuccessful cardiopulmonary resuscitation: a prospective clinical trial, by
Bottiger et al., The Lancet, 2001.
}
\usage{
cpr
}
\description{
These patients were randomly divided into a treatment group where they
received a blood thinner or the control group where they did not receive a
blood thinner. The outcome variable of interest was whether the patients
survived for at least 24 hours.
}
\examples{

table(cpr)
}
\keyword{datasets}
