% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-credits.R
\docType{data}
\name{credits}
\alias{credits}
\title{College credits.}
\format{
A data frame with 100 observations on the following variable.
\describe{
\item{credits}{Number of credits.}
}
}
\source{
Simulated data.
}
\usage{
credits
}
\description{
A simulated dataset of number of credits taken by college students each
semester.
}
\examples{

library(ggplot2)

ggplot(credits, aes(x = credits)) +
  geom_histogram(binwidth = 1)
}
\keyword{college}
\keyword{credits}
\keyword{datasets}
