% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dream.R
\docType{data}
\name{dream}
\alias{dream}
\title{Survey on views of the DREAM Act}
\format{
A data frame with 910 observations on the following 2 variables.
\describe{
\item{ideology}{a factor with levels \code{Conservative} \code{Liberal} \code{Moderate}}
\item{stance}{a factor with levels \code{No} \code{Not sure} \code{Yes}}
}
}
\source{
SurveyUSA, News Poll #18927, data collected Jan 27-29, 2012.
}
\usage{
dream
}
\description{
A SurveyUSA poll.
}
\examples{

table(dream)

}
\keyword{datasets}
