% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-drone_blades.R
\docType{data}
\name{drone_blades}
\alias{drone_blades}
\title{Quadcopter Drone Blades}
\format{
A data frame with 2000 observations on the following 2 variables.
\describe{
\item{supplier}{The supplier for the blade.}
\item{inspection}{The inspection conclusion.}
}
}
\usage{
drone_blades
}
\description{
Quality control data set for quadcopter drone blades, where this data has
been made up for an example.
}
\examples{

library(dplyr)

drone_blades \%>\%
  count(supplier, inspection)
}
\references{
OpenIntro Statistics, Third Edition and Fourth Edition.
}
\keyword{datasets}
