% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-exams.R
\docType{data}
\name{exams}
\alias{exams}
\title{Exam scores}
\format{
A data frame with 19 observations on the following variable.
\describe{\item{scores}{a numeric vector} }
}
\usage{
exams
}
\description{
Exam scores from a class of 19 students.
}
\examples{

hist(exams$scores)
}
\keyword{datasets}
