% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fact-opinion.R
\docType{data}
\name{fact_opinion}
\alias{fact_opinion}
\title{Can Americans categorize facts and opinions?}
\format{
A data frame with 5035 rows and 3 variables.
\describe{
\item{age_group}{Age group of survey participant.}
\item{fact_correct}{Number of factual statements classified correctly (out of 5).}
\item{opinion_correct}{Number of opinion statements classified correctly (out of 5).}
}
}
\source{
\href{https://www.pewresearch.org/fact-tank/2018/10/23/younger-americans-are-better-than-older-americans-at-telling-factual-news-statements-from-opinions/}{Pew Research Center}
}
\usage{
fact_opinion
}
\description{
Pew Research Center conducted a survey in 2018, asking
participants to categorize five factual and five opinion statements.
}
\examples{
boxplot(fact_correct~age_group, data = fact_opinion,
  main = "Number of Factual Statements Correctly Categorized",
  xlab = "Age Group", ylab = "Number Correct (out of 5)")

}
\keyword{datasets}
