% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-family_college.R
\docType{data}
\name{family_college}
\alias{family_college}
\title{Simulated sample of parent / teen college attendance}
\format{
A data frame with 792 observations on the following 2 variables.
\describe{
\item{teen}{Whether the teen goes to \code{college} or \code{not}.}
\item{parents}{Whether the parent holds a college \code{degree} or \code{not}.}
}
}
\source{
Simulation based off of summary information provided at
\url{https://eric.ed.gov/?id=ED460660}.
}
\usage{
family_college
}
\description{
A simulated dataset based on real population summaries.
}
\examples{

library(dplyr)

family_college |>
  count(teen, parents)
}
\keyword{datasets}
