% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fcid.R
\docType{data}
\name{fcid}
\alias{fcid}
\title{Summary of male heights from USDA Food Commodity Intake Database}
\format{
A data frame with 100 observations on the following 2 variables.
\describe{\item{height}{a numeric vector}
\item{num_of_adults}{a numeric vector} }
}
\usage{
fcid
}
\description{
Sample of heights based on the weighted sample in the survey.
}
\examples{

fcid

}
\keyword{datasets}
