\name{fheights}
\alias{fheights}
\docType{data}
\title{Female college student heights, in inches}
\description{24 sample observations.}
\usage{data("fheights")}
\format{
  A data frame with 24 observations on the following variable.
  \describe{
    \item{\code{heights}}{height, in inches}
  }
}
\examples{
data(fheights)
hist(fheights$heights)
}
\keyword{datasets}
