% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gpa.R
\docType{data}
\name{gpa}
\alias{gpa}
\title{Survey of Duke students on GPA, studying, and more}
\format{
A data frame with 55 observations on the following 5 variables.
\describe{\item{gpa}{a numeric vector} \item{studyweek}{a
numeric vector} \item{sleepnight}{a numeric vector}
\item{out}{a numeric vector} \item{gender}{a factor with
levels \code{female} \code{male}} }
}
\usage{
gpa
}
\description{
A survey of 55 Duke University students asked about their GPA, number of
hours they study at night, number of nights they go out, and their gender.
}
\examples{

gpa
}
\keyword{datasets}
