% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gss2010.R
\docType{data}
\name{gss2010}
\alias{gss2010}
\title{2010 General Social Survey}
\format{
A data frame with 2044 observations on the following 5 variables.
\describe{
\item{hrsrelax}{After an average work day, about how many hours do you have to relax or pursue activities that you enjoy}
\item{mntlhlth}{For how many days during the past 30 days was your mental health, which includes stress, depression, and problems with emotions, not good?}
\item{hrs1}{Hours worked each week.}
\item{degree}{Educational attainment or degree.}
\item{grass}{Do you think the use of marijuana should be made legal, or not?}
}
}
\source{
US 2010 General Social Survey.
}
\usage{
gss2010
}
\description{
Data from the 2010 General Social Survey.
}
\examples{


gss2010
}
\keyword{datasets}
