% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-housing.R
\docType{data}
\name{housing}
\alias{housing}
\title{Simulated data set on student housing}
\format{
A data frame with 75 observations on the following variable.
\describe{\item{cost}{a numeric vector} }
}
\usage{
housing
}
\description{
Each observation represents a simulated rent price for a student.
}
\examples{


housing
}
\keyword{datasets}
