% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ipod.R
\docType{data}
\name{ipod}
\alias{ipod}
\title{Length of songs on an iPod}
\format{
A data frame with 3000 observations on the following variable.
\describe{
\item{song_length}{Length of song (in minutes).}
}
}
\source{
Simulated data.
}
\usage{
ipod
}
\description{
A simulated data set on lengths of songs on an iPod.
}
\examples{

library(ggplot2)

ggplot(ipod, aes(x = song_length)) +
  geom_histogram(binwidth = 0.5)
}
\keyword{datasets}
\keyword{music}
