% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-jury.R
\docType{data}
\name{jury}
\alias{jury}
\title{Simulated juror dataset}
\format{
A data frame with 275 observations on the following variable.
\describe{\item{race}{a factor with levels \code{black} \code{hispanic}
\code{other} \code{white}} }
}
\usage{
jury
}
\description{
Simulated dataset of registered voters proportions and representation on
juries.
}
\examples{


jury
}
\keyword{datasets}
