% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-male_heights.R
\docType{data}
\name{male_heights}
\alias{male_heights}
\title{Sample of 100 male heights}
\format{
A data frame with 100 observations on the following variable.
\describe{\item{heights}{a numeric vector} }
}
\usage{
male_heights
}
\description{
Random sample based on Food Commodity Intake Database distribution
}
\examples{

male_heights
}
\references{
What We Eat In America - Food Commodity Intake Database.
Available at \url{https://fcid.foodrisk.org/}.
}
\keyword{datasets}
