% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-male_heights_fcid.R
\docType{data}
\name{male_heights_fcid}
\alias{male_heights_fcid}
\title{Random sample of adult male heights}
\format{
A data frame with 100 observations on the following variable.
\describe{
  \item{height_inch}{Height, in inches.}
  }
}
\source{
Simulated based on data from USDA.
}
\usage{
male_heights_fcid
}
\description{
This sample is based on data from the USDA Food Commodity Intake Database.
}
\examples{

data(male_heights_fcid)
histPlot(male_heights_fcid$height_inch)

}
\keyword{datasets}
