% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-midterms_house.R
\docType{data}
\name{midterms_house}
\alias{midterms_house}
\title{President's party performance and unemployment rate}
\format{
A data frame with 29 observations on the following 5 variables.
\describe{
  \item{year}{a numeric vector}
  \item{potus}{The president in office.}
  \item{party}{President's party.}
  \item{unemp}{Unemployment rate.}
  \item{house_change}{Change in House seats for the president's party.}
  }
}
\source{
Wikipedia.
}
\usage{
midterms_house
}
\description{
Covers midterm elections.
}
\details{
An older version of this data is at \code{\link{unemploy_pres}}.
}
\examples{

library(ggplot2)

ggplot(midterms_house, aes(x = unemp, y = house_change)) +
  geom_point()

}
\keyword{datasets}
