% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-migraine.R
\docType{data}
\name{migraine}
\alias{migraine}
\title{Migraines and acupuncture}
\format{
A data frame with 89 observations on the following 2 variables.
\describe{\item{group}{a factor with levels \code{control} \code{treatment}}
\item{pain_free}{a factor with levels \code{no} \code{yes}} }
}
\source{
G. Allais et al. Ear acupuncture in the treatment of migraine
attacks: a randomized trial on the efficacy of appropriate versus
inappropriate acupoints. In: Neurological Sci. 32.1 (2011), pp. 173-175.
}
\usage{
migraine
}
\description{
Experiment involving acupuncture and sham acupuncture (as placebo) in the
treatment of migraines.
}
\examples{


migraine


}
\keyword{datasets}
