% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openintro-package.R
\docType{package}
\name{openintro-package}
\alias{openintro}
\alias{openintro-package}
\title{openintro: Data Sets and Supplemental Functions from 'OpenIntro' Textbooks and Labs}
\description{
Supplemental functions and data for 'OpenIntro' resources, which 
    includes open-source textbooks and resources for introductory statistics at 
    openintro.org. The package contains data sets used in our open-source textbooks 
    along with custom plotting functions for reproducing book figures. Note that 
    many functions and examples include color transparency; some plotting elements 
    may not show up properly (or at all) when run in some versions of Windows 
    operating system.
}
\author{
\strong{Maintainer}: Mine Çetinkaya-Rundel \email{mine@rstudio.com} (0000-0001-6452-2420)

Authors:
\itemize{
  \item David Diez \email{david.m.diez@gmail.com}
  \item Andrew Bray \email{abray@reed.edu}
  \item Christopher Barr
}

}
\keyword{internal}
