% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openintro-package.R
\docType{package}
\name{openintro-package}
\alias{openintro}
\alias{openintro-package}
\title{openintro: Data Sets and Supplemental Functions from 'OpenIntro' Textbooks and Labs}
\description{
Supplemental functions and data for 'OpenIntro' resources, which 
    includes open-source textbooks and resources for introductory statistics 
    (<https://www.openintro.org/>). The package contains data sets used in our 
    open-source textbooks along with custom plotting functions for reproducing 
    book figures. Note that many functions and examples include color 
    transparency; some plotting elements may not show up properly (or at all) 
    when run in some versions of Windows operating system.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OpenIntroStat/openintro}
  \item Report bugs at \url{https://github.com/OpenIntroStat/openintro/issues}
}

}
\author{
\strong{Maintainer}: Mine Çetinkaya-Rundel \email{cetinkaya.mine@gmail.com} (\href{https://orcid.org/0000-0001-6452-2420}{ORCID})

Authors:
\itemize{
  \item David Diez \email{david.m.diez@gmail.com}
  \item Andrew Bray \email{andrew.bray@gmail.com}
  \item Albert Kim \email{albert.ys.kim@gmail.com}
  \item Ben Baumer \email{ben.baumer@gmail.com}
  \item Chester Ismay \email{chester.ismay@gmail.com}
  \item Christopher Barr
}

}
\keyword{internal}
