% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-present.R
\docType{data}
\name{present}
\alias{present}
\title{Birth counts}
\format{
A data frame with 63 observations on the following 3 variables.
\describe{
\item{year}{Year.}
\item{boys}{Number of boys born.}
\item{girls}{Number of girls born.}
}
}
\source{
Mathews, T. J., and Brady E. Hamilton. "Trend analysis of the sex
ratio at birth in the United States." National vital statistics reports
53.20 (2005): 1-17.
}
\usage{
present
}
\description{
An updated version of the historical Arbuthnot dataset. Numbers of boys and
girls born in the United States between 1940 and 2002.
}
\examples{

library(ggplot2)

ggplot(present, mapping = aes(x = year, y = boys / girls)) +
  geom_line()
}
\keyword{datasets}
