% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-president.R
\docType{data}
\name{president}
\alias{president}
\title{United States Presidental History}
\format{
A data frame with 67 observations on the following 5 variables.
\describe{\item{potus}{President of the United States}
\item{party}{Political party of the president} \item{start}{Start year}
\item{end}{End year} \item{vpotus}{Vice President of the United States} }
}
\source{
Presidents of the United States (table) -- infoplease.com (visited:
Nov 2nd, 2010)

\url{http://www.infoplease.com/ce6/history/A0840075.html}
}
\usage{
president
}
\description{
Summary of the changes in the president and vice president for the United
States of America.
}
\examples{


president


}
\keyword{datasets}
