\name{prison}
\alias{prison}
\docType{data}
\title{
  Prison isolation experiment
}
\description{
  Subjects from Central Prison in Raleigh, NC, volunteered for
  an experiment involving an "isolation" experience. The goal
  of the experiment was to find a treatment that reduces subjects'
  psychopathic deviant T scores. This score measures a person's
  need for control or their rebellion against control, and it is part
  of a commonly used mental health test called the Minnesota
  Multiphasic Personality Inventory (MMPI) test. 

}
\usage{data("prison")}
\format{
  A data frame with 14 observations on the following 6 variables.
  \describe{
    \item{\code{PreTrt1}}{Pre-treatment 1.}
    \item{\code{PostTrt1}}{Post-treatment 1.}
    \item{\code{PreTrt2}}{Pre-treatment 2.}
    \item{\code{PostTrt2}}{Post-treatment 2.}
    \item{\code{PreTrt3}}{Pre-treatment 3.}
    \item{\code{PostTrt3}}{Post-treatment 3.}
  }
}
\source{
  \url{http://stat.duke.edu/resources/datasets/prison-isolation}
}
\examples{
data(prison)
}
\keyword{datasets}
