% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-simulated_dist.R
\docType{data}
\name{simulated_dist}
\alias{simulated_dist}
\title{Simulated data sets, not necessarily drawn from a normal distribution.}
\format{
The format is: List of 4 $ d1: data set of 100 observations.  $ d2:
data set of 50 observations.  $ d3: num data set of 500 observations.  $ d4:
data set of 15 observations.  $ d5: num data set of 25 observations.  $ d6:
data set of 50 observations.
}
\usage{
simulated_dist
}
\description{
Data were simulated in R, and some of the simulations do not represent data
from actual normal distributions.
}
\examples{

data(simulated_dist)
lapply(simulated_dist, qqnorm)

}
\keyword{datasets}
