% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-simulated_normal.R
\docType{data}
\name{simulated_normal}
\alias{simulated_normal}
\title{Simulated data sets, drawn from a normal distribution.}
\format{
The format is: List of 3 $ n40 : 40 observations from a standard
normal distribution.  $ n100: 100 observations from a standard normal
distribution.  $ n400: 400 observations from a standard normal distribution.
}
\usage{
simulated_normal
}
\description{
Data were simulated using \code{\link{rnorm}}.
}
\examples{

data(simulated_normal)
par(mfrow = c(1, 3))
lapply(simulated_normal, qqnorm)

}
\keyword{datasets}
