% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-simulated_scatter.R
\docType{data}
\name{simulated_scatter}
\alias{simulated_scatter}
\title{Simulated data for sample scatterplots}
\format{
A data frame with 500 observations on the following 3 variables.
\describe{
\item{group}{Group, representing data for a specific plot.}
\item{x}{x-value.}
\item{y}{y-value.}
}
}
\usage{
simulated_scatter
}
\description{
Fake data.
}
\examples{

library(ggplot2)

ggplot(simulated_scatter, aes(x = x, y = y)) +
  geom_point() +
  facet_wrap(~group)
}
\keyword{datasets}
