% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sinusitis.R
\docType{data}
\name{sinusitis}
\alias{sinusitis}
\title{Sinusitis and antibiotic experiment}
\format{
A data frame with 166 observations on the following 2 variables.
\describe{
\item{group}{a factor with levels \code{control} and  \code{treatment}}
\item{self_reported_improvement}{a factor with levels \code{no} and \code{yes}}
}
}
\source{
J.M. Garbutt et al. Amoxicillin for Acute Rhinosinusitis: A
Randomized Controlled Trial. In: JAMA: The Journal of the American Medical
Association 307.7 (2012), pp. 685-692.
}
\usage{
sinusitis
}
\description{
Researchers studying the effect of antibiotic treatment for acute sinusitis
to one of two groups: treatment or control.
}
\examples{

sinusitis
}
\keyword{datasets}
