% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-starbucks.R
\docType{data}
\name{starbucks}
\alias{starbucks}
\title{Starbucks nutrition}
\format{
A data frame with 77 observations on the following 7 variables.
\describe{
  \item{item}{Food item.}
  \item{calories}{Calories.}
  \item{fat}{a numeric vector}
  \item{carb}{a numeric vector}
  \item{fiber}{a numeric vector}
  \item{protein}{a numeric vector}
  \item{type}{a factor with levels \code{bakery}, \code{bistro box}, \code{hot breakfast}, \code{parfait}, \code{petite}, \code{salad}, and \code{sandwich}}
  }
}
\source{
\url{https://www.starbucks.com/menu}, retrieved 2011-03-10.
}
\usage{
starbucks
}
\description{
Nutrition facts for several Starbucks food items
}
\examples{

starbucks

}
\keyword{datasets}
