% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-stocks_18.R
\docType{data}
\name{stocks_18}
\alias{stocks_18}
\title{Monthly Returns for a few stocks}
\format{
A data frame with 36 observations on the following 3 variables.
\describe{
\item{date}{First day of the month corresponding to the returns.}
\item{goog}{Google stock price change.}
\item{cat}{Caterpillar stock price change.}
\item{xom}{Exxon Mobil stock price change.}
}
}
\source{
Yahoo! Finance, direct download.
}
\usage{
stocks_18
}
\description{
Monthly return data for a few stocks, which covers stock prices from
November 2015 through October 2018.
}
\examples{

d <- stocks_18
dim(d)
apply(d[, 2:3], 2, mean)
apply(d[, 2:3], 2, sd)

}
\keyword{datasets}
