% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sulphinpyrazone.R
\docType{data}
\name{sulphinpyrazone}
\alias{sulphinpyrazone}
\title{Treating heart attacks}
\format{
A data frame with 1475 observations on the following 2 variables.
\describe{
\item{group}{a factor with levels \code{control} \code{treatment}}
\item{outcome}{a factor with levels \code{died} \code{lived}}
}
}
\source{
Anturane Reinfarction Trial Research Group. 1980. Sulfinpyrazone in
the prevention of sudden death after myocardial infarction. New England
Journal of Medicine 302(5):250-256.
}
\usage{
sulphinpyrazone
}
\description{
Experiment data for studying the efficacy of treating patients who have had
a heart attack with Sulphinpyrazone.
}
\examples{

sulphinpyrazone
}
\keyword{datasets}
