% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-toohey.R
\docType{data}
\name{toohey}
\alias{toohey}
\title{Simulated polling data set}
\format{
A data frame with 500 observations on the following variable.
\describe{
\item{vote_for}{a factor with levels \code{no} \code{yes}}
}
}
\usage{
toohey
}
\description{
Simulated data for a fake political candidate.
}
\examples{

toohey
}
\keyword{datasets}
