% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tourism.R
\docType{data}
\name{tourism}
\alias{tourism}
\title{Turkey tourism}
\format{
A data frame with 47 observations on the following 3 variables.
\describe{
\item{year}{a numeric vector}
\item{visitor_count_tho}{a numeric vector}
\item{tourist_spending}{a numeric vector}
}
}
\source{
Association of Turkish Travel Agencies, Foreign Visitors Figure &
Tourist Spendings By Years.
\url{http://www.tursab.org.tr/en/statistics/foreign-visitors-figure-tourist-spendings-by-years_1083.html}
}
\usage{
tourism
}
\description{
Summary of tourism in Turkey.
}
\examples{

tourism
}
\keyword{datasets}
