% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-toy_anova.R
\docType{data}
\name{toy_anova}
\alias{toy_anova}
\title{Simulated dataset for ANOVA}
\format{
A data frame with 70 observations on the following 3 variables.
\describe{
\item{group}{a factor with levels \code{I} \code{II} \code{III}}
\item{outcome}{a numeric vector}
}
}
\usage{
toy_anova
}
\description{
Simulated dataset for getting a better understanding of intuition that
ANOVA is based off of.
}
\examples{

toy_anova
}
\keyword{datasets}
