\name{transplant}
\alias{transplant}
\docType{data}
\title{Transplant consultant success rate (fake data)}
\description{Summarizing whether there was or was not a complication for 62 patients who used a particular medical consultant.}
\usage{data("transplant")}
\format{
  A data frame with 62 observations on the following variable.
  \describe{
    \item{\code{outcome}}{a factor with levels \code{complications} \code{okay}}
  }
}
\examples{
data(transplant)
## maybe str(transplant) ; plot(transplant) ...
}
\keyword{datasets}