% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-xom.R
\docType{data}
\name{xom}
\alias{xom}
\title{Exxon Mobile stock data}
\format{
A data frame with 98 observations on the following 7 variables.
\describe{
\item{date}{Date.}
\item{open}{a numeric vector}
\item{high}{a numeric vector}
\item{low}{a numeric vector}
\item{close}{a numeric vector}
\item{volume}{a numeric vector}
\item{adj_close}{a numeric vector}
}
}
\source{
Yahoo! Finance.
}
\usage{
xom
}
\description{
Monthly data covering 2006 through early 2014.
}
\examples{

xom
}
\keyword{datasets}
